Move identity token button
